Abbreviated Pagination
Added an abbreviated pagination above the pictures. Extended pagination is now below the pictures. 
If you are on page one the abbreviated pagination shows:
"1 2 3 > >>"
As usual, the single arrow takes you to the next page, the double arrow takes you to the last page.
If you are on page five it would show as:
"<< < 3 4 5 6 7 > >>"
And on the last page (here 10) :
"<< < 8 9 10"